+++ src/main/java/kitchenpos/menu/infra/JpaMenuRepository.java
@@ -2,8 +2,8 @@ package kitchenpos.menu.infra;
 
 import java.util.List;
 import java.util.UUID;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.Menu;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
