+++ src/main/java/kitchenpos/menu/tobe/domain/MenuRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.tobe.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menu.tobe.domain.Menu;
 
 public interface MenuRepository {
