+++ src/main/java/kitchenpos/order/deliveryorder/application/DeliveryOrderService.java
@@ -9,8 +9,8 @@ import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import kitchenpos.deliveryagency.DeliveryAgencyClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.Menu;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrder;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrderLineItem;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrderRepository;
