+++ src/main/java/kitchenpos/order/deliveryorder/domain/DeliveryOrderLineItem.java
@@ -4,7 +4,6 @@ import java.math.BigDecimal;
 import java.util.UUID;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.ForeignKey;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -12,7 +11,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import javax.persistence.Transient;
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.tobe.domain.Menu;
 
 @Table(name = "delivery_order_line_item")
 @Entity
