+++ src/main/java/kitchenpos/order/eatinorder/application/EatInOrderService.java
@@ -8,8 +8,8 @@ import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import kitchenpos.deliveryagency.DeliveryAgencyClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.Menu;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import kitchenpos.order.eatinorder.domain.EatInOrder;
 import kitchenpos.order.eatinorder.domain.EatInOrderLineItem;
 import kitchenpos.order.eatinorder.domain.EatInOrderRepository;
