+++ src/main/java/kitchenpos/product/tobe/application/ProductService.java
@@ -3,8 +3,8 @@ package kitchenpos.product.tobe.application;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.Menu;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import kitchenpos.product.tobe.application.dto.ChangeProductPriceRequest;
 import kitchenpos.product.tobe.application.dto.ChangeProductPriceResponse;
 import kitchenpos.product.tobe.application.dto.CreateProductRequest;
