+++ src/test/java/kitchenpos/application/DeliveryOrderServiceTest.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Random;
 import java.util.UUID;
 import kitchenpos.Fixtures;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import kitchenpos.order.deliveryorder.application.DeliveryOrderService;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrder;
 import kitchenpos.order.deliveryorder.domain.DeliveryOrderLineItem;
