+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -10,8 +10,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.tobe.domain.Menu;
+import kitchenpos.menu.tobe.domain.MenuRepository;
 import kitchenpos.product.tobe.application.ProductService;
 import kitchenpos.product.tobe.application.dto.ChangeProductPriceRequest;
 import kitchenpos.product.tobe.application.dto.CreateProductRequest;
