+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -3,6 +3,7 @@ package kitchenpos.menu.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import kitchenpos.Fixtures;
+import kitchenpos.menu.tobe.domain.Menu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.CsvSource;
